+++ Source/OS9exec_core/modstuff.h
@@ -85,7 +85,7 @@ void init_exceptions(ushort pid);
 void init_traphandlers(ushort pid);
 
 ushort calc_parity(ushort *p, ushort numwords);
-ulong  calc_crc(byte *p, ulong size, ulong accum);
+uint32_t  calc_crc(byte *p, uint32_t size, uint32_t accum);
 void   mod_crc(mod_exec *m);
 
 os9err prepData(ushort    pid,
